confd: move srx_module to src/lib
