+++ src/confd/src/confd/Makefile.am
@@ -7,6 +7,6 @@ plugin_LTLIBRARIES      = confd-plugin.la
 confd_plugin_la_CFLAGS  = $(augeas_CFLAGS) $(jansson_CFLAGS) $(libite_CFLAGS) $(sysrepo_CFLAGS) $(CFLAGS)
 confd_plugin_la_LIBADD  = $(augeas_LIBS)   $(jansson_LIBS)   $(libite_LIBS)   $(sysrepo_LIBS)
 confd_plugin_la_LDFLAGS = -module -avoid-version -shared
-confd_plugin_la_SOURCES = core.c core.h helpers.c helpers.h srx_module.c srx_module.h srx_val.c srx_val.h \
+confd_plugin_la_SOURCES = core.c core.h helpers.c helpers.h srx_val.c srx_val.h \
 			  ietf-system.c ietf-interfaces.c systemv.c systemv.h ../lib/lyx.c ../lib/lyx.h dagger.c dagger.h \
-			  infix-dhcp.c
+			  infix-dhcp.c ../lib/srx_module.c ../lib/srx_module.h
