+++ src/confd/src/confd/core.c
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 
 #include "core.h"
-#include "srx_module.h"
+#include "../lib/srx_module.h"
 
 static struct confd confd;
